requirements/common.txt
Django>=2.2,<3
django-environ

psycopg2

celery

ipython

django-templated-email

sentry-sdk

newrelic
djangorestframework>=3.10
git+https://github.com/razortheory/drf-secure-token

django-fsm
django-fsm-admin
fcm-django>=0.3.2
